Structured
Structured the project to look more appealing. Also configured addtreetype to work.
