+++ ImplementationFX/model/ScoutModels/Scout.java
@@ -1,10 +1,11 @@
-package model;
+package model.ScoutModels;
 
 import database.*;
 import exception.*;
 import impresario.*;
 import common.*;
 import event.*;
+import model.EntityBase;
 
 import java.sql.SQLException;
 import java.text.SimpleDateFormat;
