+++ ImplementationFX/model/TreeModels/TreeCollection.java
@@ -1,8 +1,8 @@
-package model;
+package model.TreeModels;
 
-import database.*;
-import exception.*;
 import impresario.*;
+import model.EntityBase;
+
 import java.util.Properties;
 import java.util.Vector;
