+++ ImplementationFX/userinterface/TreeViews/AddTreeView.java
@@ -1,4 +1,4 @@
-package userinterface;
+package userinterface.TreeViews;
 
 import impresario.IModel;
 import javafx.geometry.Insets;
@@ -10,8 +10,10 @@ import javafx.scene.layout.*;
 import javafx.scene.text.Font;
 import javafx.scene.text.FontWeight;
 import javafx.scene.text.Text;
-import model.TreeTypeCollection;
-import model.TreeType;
+import model.TreeTypeModels.TreeTypeCollection;
+import model.TreeTypeModels.TreeType;
+import userinterface.MessageView;
+import userinterface.View;
 
 import java.text.SimpleDateFormat;
 import java.util.Date;
